agregando datos

+++ app/Models/Plex.php
@@ -845,15 +845,18 @@ class Plex {
         }
 
 
-        $opts = [
-            "http" => [
-                "method" => "DELETE",
-                "header" => "X-Plex-Token: ".$owner['user']['authToken']
-            ]
-        ];
+        if(count($data) > 0){
+            
+            $opts = [
+                "http" => [
+                    "method" => "DELETE",
+                    "header" => "X-Plex-Token: ".$owner['user']['authToken']
+                ]
+            ];
 
-        $context = stream_context_create($opts);
-        $response = file_get_contents('https://clients.plex.tv/api/v2/sharings/'.$data['id'], false, $context);
-        $data = simplexml_load_string($response);
+            $context = stream_context_create($opts);
+            $response = file_get_contents('https://clients.plex.tv/api/v2/sharings/'.$data['id'], false, $context);
+            $data = simplexml_load_string($response);
+        }
     }
 }
